Initial Commit w/ Makefile bootsector, uefi exit
Makefile runs on OSX, main changes for linux include changing the commands for
loopback devices, creating a FAT filesystem, and executing the mount commands as
root. The bootsector simply prints an error message for booting in legacy BIOS
mode and then hangs, this should likely eventually prompt to restart the system.
Finally, once booted with uefi, the memory map is printed, and then exits boot
services.