Reorganized source tree, initialized gdt, exited uefi, created console.
I did a lot in this commit, and I _need_ to start using git more, but I'll try
to describe the highlights here. First I created a few folders to reorganize
the source tree, mainly this includes `src` for the source code, `include` for
include files, and `skel` for the template of the ESP image.

Next I exited UEFI and started printing text to the screen and implemented a
basic printf. I also created a GDT and I am in the midst of creating an IDT and
TSS, but I think I want to get memory management working first. I ran into a
bit of a chicken vs egg problem where I wanted to get and IDT up first so I
could easily debug page faults, but for that I need a TSS which requires a
stack which I need to allocate.

I also moved to clang and nasm here because I was frustrated with the gnu
tools. In the future, I'll probably remove the nasm invocation and switch
entirely to clang, just to keep it all consistent.

This commit also includes the binary font file within the `skel` directory, I
know generally binary files shouldn't be included with git, but I think this is
a valid exception.