Inital memory management and updated build system
This commit includes a physical and virtual page management as well as a basic
heap. In order to properly get this all working, I switched back to using gnu
ld as my linker because I wanted to have linker scripts avaliable to add in
arbitary labels. I also relocated the kernel up to the last gigabyte in memory
and gave it a new stack, removing all reliance on memory given by UEFI Boot
Services (except the TSS & IDT which should be updated in the next commit).

Several minor changes also occured, namely the introduction of a kernel panic,
which should later be expanded to include a register dump, the build system now
detects all sources automatically, and the previous stall.c was renamed to a
more fitting bootstrap.c.