Prepared for running init process
I created a simple simple init program (which currently just hangs), such that
I can use it while building the init program loader. Currently it is loaded
into memory by uefi and is treated as bootstrap data.

This commit also includes a few edits that should have been in previous
commits, mainly involving symbol renaming for the beginning of kernel sections,
as well as updating the prototype for map_page.