Changed virtual memory layout
With this commit, there is now a direct map of physical memory directly
following the cannonical address hole. The general heap has also been moved in
order to make space for the address space and capability heaps which which will
be added later.

The kernel has lost safety with this commit, as it now executes within that
direct map of physical memory which is read, write, execute for all memory.
Thus kernel code can be written to (by the kernel only), and the kernel data
and bss sections can also be executed.